WIP-upgrade node and coding style
Convert from callback and promise to asyc/await, upgrade all the node module packages to the latest version
