--- .eslintrc
@@ -1,3 +1,0 @@
-"parserOptions": {
-   "ecmaVersion": 6
- }
