+++ .eslintrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+  "parserOptions": {
+    "ecmaVersion": 12,
+    "env": {
+      "node": true,
+      "es2021": true
+    }
+  }
+}
