+++ config/default.js
@@ -1,6 +1,7 @@
 /**
  * The configuration file.
  */
+require('dotenv').config()
 module.exports = {
   LOG_LEVEL: process.env.LOG_LEVEL,
   PORT: process.env.PORT,
@@ -34,7 +35,7 @@ module.exports = {
   TEMPLATE_MAP: process.env.TEMPLATE_MAP,
   SENDGRID_API_KEY: process.env.SENDGRID_API_KEY || '',
   EMAIL_FROM: process.env.EMAIL_FROM || 'no-reply@topcoder.com',
-  
+
   // temporary not in use feature
   EMAIL_MAX_ERRORS: process.env.EMAIL_MAX_ERRORS || 2,
   EMAIL_PAUSE_TIME: process.env.EMAIL_PAUSE_TIME || 30,
@@ -42,7 +43,7 @@ module.exports = {
   //in every 2 minutes will retry for failed status
   EMAIL_RETRY_SCHEDULE: process.env.EMAIL_RETRY_SCHEDULE || '0 */2 * * * *',
   //wont't retry failed emails older than this time (msec)
-  EMAIL_RETRY_MAX_AGE: process.env.EMAIL_RETRY_MAX_AGE || 1000*60*60*24,
+  EMAIL_RETRY_MAX_AGE: process.env.EMAIL_RETRY_MAX_AGE || 1000 * 60 * 60 * 24,
 
   API_CONTEXT_PATH: process.env.API_CONTEXT_PATH || '/v5/email',
