+++ connect/service.js
@@ -1,7 +1,6 @@
 /**
  * This is TopCoder connect email service.
  */
-'use strict';
 
 const sgMail = require('@sendgrid/mail');
 const config = require('config');
@@ -9,18 +8,19 @@ const config = require('config');
 // set api key for SendGrid email client
 sgMail.setApiKey(config.SENDGRID_API_KEY);
 
-const sendEmail = (templateId, message) => { // send email
+const sendEmail = async (templateId, message) => { // send email
 
   const from = message.from ? message.from : config.EMAIL_FROM;
   const replyTo = message.replyTo ? message.replyTo : config.EMAIL_FROM;
-  const substitutions = message.data ;
-  const categories = message.categories ? message.categories: [];
+  const substitutions = message.data;
+  const categories = message.categories ? message.categories : [];
   const to = message.recipients;
-  const cc =  message.cc ? message.cc : [];
+  const cc = message.cc ? message.cc : [];
   const bcc = message.bcc ? message.bcc : [];
-  
-  if (message.version && message.version=="v3"){
-    return  sgMail.send({
+
+  let msg = {}
+  if (message.version && message.version == "v3") {
+    msg = ({
       to,
       templateId,
       dynamicTemplateData: substitutions,
@@ -30,8 +30,8 @@ const sendEmail = (templateId, message) => { // send email
       cc,
       bcc,
     });
-  } else{
-    return  sgMail.send({
+  } else {
+    msg = ({
       to,
       templateId,
       substitutions,
@@ -43,6 +43,7 @@ const sendEmail = (templateId, message) => { // send email
       bcc,
     });
   }
+  return await sgMail.send(msg);
 }
 module.exports = {
   sendEmail,
