+++ index.js
@@ -1,13 +1,15 @@
-/**
- * This is entry point of the TopCoder notification server module.
- */
-'use strict';
-
+require('dotenv').config()
 const config = require('config');
+const jwtAuth = require('tc-core-library-js').middleware.jwtAuthenticator;
+const express = require('express');
 const _ = require('lodash');
 const schedule = require('node-schedule');
+const cors = require('cors');
+const bodyParser = require('body-parser');
+const helper = require('./src/common/helper');
 const logger = require('./src/common/logger');
 const errors = require('./src/common/errors');
+const { initServer } = require('./src/init');
 
 config.TEMPLATE_MAP = JSON.parse(config.TEMPLATE_MAP);
 
@@ -35,21 +37,6 @@ function setConfig(cfg) {
 }
 
 /**
- * Add topic handler for topic, override existing one if any.
- * @param {String} topic the topic name
- * @param {Function} handler the handler
- */
-function addTopicHandler(topic, handler) {
-  if (!topic) {
-    throw new errors.ValidationError('Missing topic.');
-  }
-  if (!handler) {
-    throw new errors.ValidationError('Missing handler.');
-  }
-  handlers[topic] = handler;
-}
-
-/**
  * Remove topic handler for topic.
  * @param {String} topic the topic name
  */
@@ -69,28 +56,101 @@ function getAllHandlers() {
 }
 
 /**
- * Start the notification server.
+ * Add topic handler for topic, override existing one if any.
+ * @param {String} topic the topic name
+ * @param {Function} handler the handler
  */
-function start() {
-  if (_.isEmpty(handlers)) {
-    throw new errors.ValidationError('Missing handler(s).');
+function addTopicHandler(topic, handler) {
+  if (!topic) {
+    throw new errors.ValidationError('Missing topic.');
   }
-  // load app only after config is set
-  const app = require('./src/app');
-  schedule.scheduleJob(config.EMAIL_RETRY_SCHEDULE, function() {
-    app.retryEmail(handlers).catch((err) => logger.error(err));
-  });
-  return app.start(handlers).catch((err) => logger.error(err));
+  if (!handler) {
+    throw new errors.ValidationError('Missing handler.');
+  }
+  handlers[topic] = handler;
 }
 
-/**
- * Initialize database. All tables are cleared and re-created.
- * @returns {Promise} promise to init db
- */
-function initDatabase() {
-  // load models only after config is set
-  const models = require('./src/models');
-  return models.init(true);
+const app = express();
+app.set('port', config.PORT);
+
+app.use(cors());
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: true }));
+
+const apiRouter = express.Router();
+
+// load all routes
+_.each(require('./src/routes'), (verbs, url) => {
+  _.each(verbs, (def, verb) => {
+    const actions = [];
+    const method = require('./src/controllers/' + def.controller)[def.method];
+    if (!method) {
+      throw new Error(def.method + ' is undefined');
+    }
+    actions.push((req, res, next) => {
+      req.signature = `${def.controller}#${def.method}`;
+      next();
+    });
+    if (url !== '/health') {
+      actions.push(jwtAuth());
+      actions.push((req, res, next) => {
+        if (!req.authUser) {
+          return next(new errors.UnauthorizedError('Authorization failed.'));
+        }
+        req.user = req.authUser;
+        return next();
+      });
+    }
+    actions.push(method);
+    apiRouter[verb](url, helper.autoWrapExpress(actions));
+  });
+});
+
+app.use(config.API_CONTEXT_PATH, apiRouter);
+
+console.log('app.start [3]')
+app.use((req, res) => {
+  res.status(404).json({ error: 'route not found' });
+});
+
+app.use((err, req, res, next) => { // eslint-disable-line
+  logger.logFullError(err, req.signature);
+  let status = err.httpStatus || 500;
+  if (err.isJoi) {
+    status = 400;
+  }
+  // from express-jwt
+  if (err.name === 'UnauthorizedError') {
+    status = 401;
+  }
+  res.status(status);
+  if (err.isJoi) {
+    res.json({
+      error: 'Validation failed',
+      details: err.details,
+    });
+  } else {
+    res.json({
+      error: err.message,
+    });
+  }
+});
+
+
+function start() {
+
+  initServer(handlers).then(() => {
+    if (_.isEmpty(handlers)) {
+      throw new errors.ValidationError('Missing handler(s).');
+    }
+
+    schedule.scheduleJob(config.EMAIL_RETRY_SCHEDULE, function () {
+      app.retryEmail(handlers).catch((err) => logger.error(err));
+    });
+    app.listen(app.get('port'), () => {
+      logger.info(`Express server listening on port ${app.get('port')}`);
+    });
+  })
 }
 
 // Exports
@@ -99,6 +159,5 @@ module.exports = {
   addTopicHandler,
   removeTopicHandler,
   getAllHandlers,
-  start,
-  initDatabase,
+  start
 };
