--- src/bootstrap.js
@@ -1,8 +1,0 @@
-/**
- * Init app
- */
-
-'use strict';
-
-global.Promise = require('bluebird');
-const logger = require('./common/logger');
