+++ src/common/errors.js
@@ -1,7 +1,6 @@
 /**
  * This file defines application errors.
  */
-'use strict';
 
 const util = require('util');
