+++ src/controllers/HealthController.js
@@ -1,7 +1,6 @@
 /**
  * Contains endpoints related to service health
  */
-'use strict';
 
 /**
  * Health Check.
