+++ src/models/datasource.js
@@ -13,23 +13,20 @@ const config = require('config');
 const Sequelize = require('sequelize');
 const logger = require('../common/logger');
 
-Sequelize.Promise = require('bluebird');
 let sequelizeInstance = null;
 
 /**
  * get sequelize instance
  */
-function getSequelize() {
+async function getSequelize() {
   if (!sequelizeInstance) {
     sequelizeInstance = new Sequelize(config.DATABASE_URL, config.DATABASE_OPTIONS);
-    sequelizeInstance
-      .authenticate()
-      .then(() => {
-        logger.info('Database connection has been established successfully.');
-      })
-      .catch((err) => {
-        logger.error('Unable to connect to the database:', err);
-      });
+    try {
+      await sequelizeInstance.authenticate()
+      logger.info('Database connection has been established successfully.');
+    } catch (e) {
+      logger.error('Unable to connect to the database:', err);
+    }
   }
   return sequelizeInstance;
 }
