+++ src/models/index.js
@@ -9,12 +9,25 @@
  * @version     1.0
  */
 
-const sequelize = require('./datasource').getSequelize();
+const sequelizeInstance = require('./datasource');
 const DataTypes = require('sequelize/lib/data-types');
+const Email = require('./Email')
 
-const Email = require('./Email')(sequelize, DataTypes);
+async function loadSequelizeModule() {
+  return await sequelizeInstance.getSequelize();
+}
+async function loadEmailModule() {
+  const sequelize = await loadSequelizeModule();
+  return await Email(sequelize, DataTypes);
+}
+
+async function init() {
+  const sequelize = await loadSequelizeModule();
+  await sequelize.sync();
+}
+
 
 module.exports = {
-  Email,
-  init: () => sequelize.sync(),
+  Email: loadEmailModule,
+  init,
 };
