+++ src/routes.js
@@ -1,4 +1,3 @@
-'use strict';
 
 module.exports = {
   '/health': {
